update feature code for movie app

+++ composeApp/src/commonMain/kotlin/com/calvin/box/movie/feature/settings/SettingsScreen.kt
@@ -224,13 +224,13 @@ class SettingsScreen:Screen {
                     SingleChoicePreference(
                         selectedKey = dohIndex.toString(),
                         onSelectionChanged = {
-                            dohIndex = it.toUInt()
-                            viewModel.eventSink(SettingsUiEvent.SetDoh(dohList[dohIndex.toInt()]))
+                            dohIndex = it.toInt()
+                            viewModel.eventSink(SettingsUiEvent.SetDoh(dohList[dohIndex]))
 
                         },
                         title = "Doh",
                         entries = dohMap,
-                        summary = dohList[dohIndex.toInt()].name
+                        summary = dohList[dohIndex].name
                     )
                 }
                 item { PreferenceDivider() }
